[SOIN] Extrais un composant pour le tiroir de légende
