+++ svelte/lib/risques/TiroirLegendeVraisemblance.svelte
@@ -1,62 +1,35 @@
 <script lang="ts">
   import type { ReferentielVraisemblances } from './risques.d';
+  import TiroirLegende from './TiroirLegende.svelte';
 
-  export let ouvert = true;
+  export let ouvert: boolean;
   export let referentielVraisemblances: ReferentielVraisemblances;
-
-  const fermeTiroir = () => {
-    ouvert = false;
-  };
 </script>
 
-<div class="tiroir" class:ouvert>
-  <div class="entete-tiroir">
+<TiroirLegende
+  bind:ouvert
+  titre="Comprendre les 5 niveaux de vraisemblance"
+  sousTitre="Découvrez les 5 niveaux définis par EBIOS Risk manager."
+>
+  {#each Array(5) as _, index (index)}
+    {@const [id, niveauVraisemblance] = Object.entries(
+      referentielVraisemblances
+    ).filter(([_, niveau]) => niveau.position === 4 - index)[0]}
     <div>
-      <h2 class="titre-tiroir-legende">
-        Comprendre les 5 niveaux de vraisemblance
-      </h2>
+      <h4>
+        <span class={`badge-niveau niveau-vraisemblance ${id}`}
+          >{niveauVraisemblance.position}</span
+        >
+        {niveauVraisemblance.libelle}
+      </h4>
+      <span>
+        {niveauVraisemblance.description}
+      </span>
     </div>
-    <button class="fermeture" on:click={fermeTiroir}>✕</button>
-  </div>
-  <div class="contenu-legende-gravite">
-    <h3>Découvrez les 5 niveaux définis par EBIOS Risk manager.</h3>
-    {#each Array(5) as _, index (index)}
-      {@const [id, niveauVraisemblance] = Object.entries(
-        referentielVraisemblances
-      ).filter(([id, niveau]) => niveau.position === 4 - index)[0]}
-      <div>
-        <h4>
-          <span class={`badge-niveau niveau-vraisemblance ${id}`}
-            >{niveauVraisemblance.position}</span
-          >
-          {niveauVraisemblance.libelle}
-        </h4>
-        <span>
-          {niveauVraisemblance.description}
-        </span>
-      </div>
-    {/each}
-  </div>
-</div>
+  {/each}
+</TiroirLegende>
 
 <style>
-  .tiroir {
-    height: 100%;
-    min-width: 650px;
-    max-width: 650px;
-    position: fixed;
-    left: 100vw;
-    top: 0;
-    overflow-y: scroll;
-    transition: transform 0.2s ease-in-out;
-    box-shadow: -10px 0 34px 0 #00000026;
-    background: #fff;
-    visibility: hidden;
-    z-index: 20;
-    display: flex;
-    flex-direction: column;
-  }
-
   .badge-niveau {
     margin-right: 8px;
     display: inline-block;
@@ -91,49 +64,5 @@
 
   h4 {
     margin: 0 0 8px;
-  }
-
-  h3 {
-    margin: 0;
-    line-height: 1.375rem;
-    font-size: 1rem;
-  }
-
-  .ouvert {
-    transform: translateX(-100%);
-    visibility: visible;
-  }
-
-  .entete-tiroir {
-    background: #f1f5f9;
-    text-align: left;
-    padding: 32px;
-    overflow: hidden;
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-start;
-  }
-
-  .fermeture {
-    font-weight: bold;
-    background: none;
-    width: 2em;
-    height: 2em;
-    cursor: pointer;
-    border: none;
-  }
-
-  .titre-tiroir-legende {
-    font-size: 1.625rem;
-    margin: 0;
-  }
-
-  .contenu-legende-gravite {
-    text-align: left;
-    padding: 32px;
-    display: flex;
-    flex-direction: column;
-    gap: 24px;
-    flex: 1;
   }
 </style>
